chore: repo refactor and clean up
