+++ src/components/icons.tsx
@@ -1,220 +1,55 @@
-import * as React from "react";
+import type { SVGProps } from "react";
 
-export const HeartIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width={24}
-    height={24}
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      d="M12 21.35L10.55 20.03C5.4 15.36 2 12.27 2 8.5C2 5.41 4.42 3 7.5 3C9.24 3 10.91 3.81 12 5.08C13.09 3.81 14.76 3 16.5 3C19.58 3 22 5.41 22 8.5C22 12.27 18.6 15.36 13.45 20.03L12 21.35Z"
-      fill="currentColor"
-    />
-  </svg>
-);
+export function XIcon(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      width={24}
+      height={25}
+      viewBox="0 0 24 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M18.0014 2.90002H21.1954L14.2183 10.8947L22.4272 21.7754H16.0005L10.9633 15.1777L5.20606 21.7754H2.00908L9.47118 13.2214L1.59998 2.90151H8.19031L12.7366 8.93099L18.0014 2.90002ZM16.8782 19.8593H18.6485L7.22333 4.71646H5.32507L16.8782 19.8593Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
 
-export const XIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width={24}
-    height={25}
-    viewBox="0 0 24 25"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <g clipPath="url(#clip0_58_554)">
-      <mask
-        id="mask0_58_554"
-        style={{
-          maskType: "luminance",
-        }}
-        maskUnits="userSpaceOnUse"
-        x={0}
-        y={0}
-        width={24}
-        height={25}
-      >
-        <path d="M0 0.5H24V24.5H0V0.5Z" fill="white" />
-      </mask>
-      <g mask="url(#mask0_58_554)">
-        <path
-          d="M18.0014 2.90002H21.1954L14.2183 10.8947L22.4272 21.7754H16.0005L10.9633 15.1777L5.20606 21.7754H2.00908L9.47118 13.2214L1.59998 2.90151H8.19031L12.7366 8.93099L18.0014 2.90002ZM16.8782 19.8593H18.6485L7.22333 4.71646H5.32507L16.8782 19.8593Z"
-          fill="currentColor"
-        />
-      </g>
-    </g>
-    <defs>
-      <clipPath id="clip0_58_554">
-        <rect
-          width={24}
-          height={24}
-          fill="white"
-          transform="translate(0 0.5)"
-        />
-      </clipPath>
-    </defs>
-  </svg>
-);
+export function GithubIcon(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      width={24}
+      height={25}
+      viewBox="0 0 24 25"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path
+        d="M12 2.5C10.6868 2.5 9.38642 2.75866 8.17317 3.2612C6.95991 3.76375 5.85752 4.50035 4.92893 5.42893C3.05357 7.3043 2 9.84784 2 12.5C2 16.92 4.87 20.67 8.84 22C9.34 22.08 9.5 21.77 9.5 21.5V19.81C6.73 20.41 6.14 18.47 6.14 18.47C5.68 17.31 5.03 17 5.03 17C4.12 16.38 5.1 16.4 5.1 16.4C6.1 16.47 6.63 17.43 6.63 17.43C7.5 18.95 8.97 18.5 9.54 18.26C9.63 17.61 9.89 17.17 10.17 16.92C7.95 16.67 5.62 15.81 5.62 12C5.62 10.89 6 10 6.65 9.29C6.55 9.04 6.2 8 6.75 6.65C6.75 6.65 7.59 6.38 9.5 7.67C10.29 7.45 11.15 7.34 12 7.34C12.85 7.34 13.71 7.45 14.5 7.67C16.41 6.38 17.25 6.65 17.25 6.65C17.8 8 17.45 9.04 17.35 9.29C18 10 18.38 10.89 18.38 12C18.38 15.82 16.04 16.66 13.81 16.91C14.17 17.22 14.5 17.83 14.5 18.76V21.5C14.5 21.77 14.66 22.09 15.17 22C19.14 20.66 22 16.92 22 12.5C22 11.1868 21.7413 9.88642 21.2388 8.67317C20.7362 7.45991 19.9997 6.35752 19.0711 5.42893C18.1425 4.50035 17.0401 3.76375 15.8268 3.2612C14.6136 2.75866 13.3132 2.5 12 2.5Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
 
-export const GithubIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width={24}
-    height={25}
-    viewBox="0 0 24 25"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      d="M12 2.5C10.6868 2.5 9.38642 2.75866 8.17317 3.2612C6.95991 3.76375 5.85752 4.50035 4.92893 5.42893C3.05357 7.3043 2 9.84784 2 12.5C2 16.92 4.87 20.67 8.84 22C9.34 22.08 9.5 21.77 9.5 21.5V19.81C6.73 20.41 6.14 18.47 6.14 18.47C5.68 17.31 5.03 17 5.03 17C4.12 16.38 5.1 16.4 5.1 16.4C6.1 16.47 6.63 17.43 6.63 17.43C7.5 18.95 8.97 18.5 9.54 18.26C9.63 17.61 9.89 17.17 10.17 16.92C7.95 16.67 5.62 15.81 5.62 12C5.62 10.89 6 10 6.65 9.29C6.55 9.04 6.2 8 6.75 6.65C6.75 6.65 7.59 6.38 9.5 7.67C10.29 7.45 11.15 7.34 12 7.34C12.85 7.34 13.71 7.45 14.5 7.67C16.41 6.38 17.25 6.65 17.25 6.65C17.8 8 17.45 9.04 17.35 9.29C18 10 18.38 10.89 18.38 12C18.38 15.82 16.04 16.66 13.81 16.91C14.17 17.22 14.5 17.83 14.5 18.76V21.5C14.5 21.77 14.66 22.09 15.17 22C19.14 20.66 22 16.92 22 12.5C22 11.1868 21.7413 9.88642 21.2388 8.67317C20.7362 7.45991 19.9997 6.35752 19.0711 5.42893C18.1425 4.50035 17.0401 3.76375 15.8268 3.2612C14.6136 2.75866 13.3132 2.5 12 2.5Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export const LinkedInIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width={30}
-    height={31}
-    viewBox="0 0 30 31"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      d="M23.75 4.25C24.413 4.25 25.0489 4.51339 25.5178 4.98223C25.9866 5.45107 26.25 6.08696 26.25 6.75V24.25C26.25 24.913 25.9866 25.5489 25.5178 26.0178C25.0489 26.4866 24.413 26.75 23.75 26.75H6.25C5.58696 26.75 4.95107 26.4866 4.48223 26.0178C4.01339 25.5489 3.75 24.913 3.75 24.25V6.75C3.75 6.08696 4.01339 5.45107 4.48223 4.98223C4.95107 4.51339 5.58696 4.25 6.25 4.25H23.75ZM23.125 23.625V17C23.125 15.9192 22.6957 14.8828 21.9315 14.1185C21.1672 13.3543 20.1308 12.925 19.05 12.925C17.9875 12.925 16.75 13.575 16.15 14.55V13.1625H12.6625V23.625H16.15V17.4625C16.15 16.5 16.925 15.7125 17.8875 15.7125C18.3516 15.7125 18.7967 15.8969 19.1249 16.2251C19.4531 16.5533 19.6375 16.9984 19.6375 17.4625V23.625H23.125ZM8.6 11.2C9.15695 11.2 9.6911 10.9788 10.0849 10.5849C10.4788 10.1911 10.7 9.65695 10.7 9.1C10.7 7.9375 9.7625 6.9875 8.6 6.9875C8.03973 6.9875 7.50241 7.21007 7.10624 7.60624C6.71007 8.00241 6.4875 8.53973 6.4875 9.1C6.4875 10.2625 7.4375 11.2 8.6 11.2ZM10.3375 23.625V13.1625H6.875V23.625H10.3375Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export const PeaceIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      d="M11.7562 0.0656118C10.1062 -0.107826 8.63437 1.08749 8.46093 2.73749L8.15624 5.57811L7.34531 3.26249C6.79687 1.70155 5.08124 0.881237 3.52031 1.42967C1.95937 1.97811 1.13437 3.69374 1.68281 5.25467L4.15781 12.2906C3.79138 12.6403 3.49981 13.0607 3.30079 13.5265C3.10176 13.9922 2.99943 14.4935 2.99999 15V16.125C2.99999 20.475 6.52499 24 10.875 24H13.125C17.475 24 21 20.475 21 16.125V10.5C21 8.8453 19.6547 7.49999 18 7.49999C17.6297 7.49999 17.2781 7.56561 16.95 7.68749C16.4625 6.68905 15.4359 5.99999 14.25 5.99999H14.1469L14.4234 3.36092C14.5969 1.71092 13.4016 0.239049 11.7516 0.0656118H11.7562ZM11.7469 7.34999C11.4328 7.82342 11.25 8.39061 11.25 8.99999V11.3437C11.0109 11.2828 10.7578 11.25 10.5 11.25H10.1531L9.89999 10.5328L10.6969 2.97186C10.7391 2.55936 11.1094 2.25936 11.5219 2.30624C11.9344 2.35311 12.2344 2.71874 12.1875 3.13124L11.7422 7.35467L11.7469 7.34999ZM5.22187 4.01249L7.76718 11.25H6.74999C6.56249 11.25 6.37499 11.264 6.19218 11.2922L3.80624 4.50936C3.67031 4.1203 3.87656 3.68905 4.26562 3.55311C4.65468 3.41717 5.08593 3.62342 5.22187 4.01249ZM13.5 8.99999C13.5 8.58749 13.8375 8.24999 14.25 8.24999C14.6625 8.24999 15 8.58749 15 8.99999V11.25C15 11.6625 14.6625 12 14.25 12C13.8375 12 13.5 11.6625 13.5 11.25V8.99999ZM15.3 14.0625C15.7875 15.0609 16.8141 15.75 18 15.75C18.2578 15.75 18.5109 15.7172 18.75 15.6562V16.125C18.75 19.2328 16.2328 21.75 13.125 21.75H10.875C7.76718 21.75 5.24999 19.2328 5.24999 16.125V15C5.24999 14.1703 5.92031 13.5 6.74999 13.5H10.5C10.9125 13.5 11.25 13.8375 11.25 14.25C11.25 14.6625 10.9125 15 10.5 15H8.62499C8.00156 15 7.49999 15.5015 7.49999 16.125C7.49999 16.7484 8.00156 17.25 8.62499 17.25H10.5C12.1547 17.25 13.5 15.9047 13.5 14.25V14.1562C13.7391 14.2172 13.9922 14.25 14.25 14.25C14.6203 14.25 14.9719 14.1844 15.3 14.0625ZM18.75 12.75C18.75 13.1625 18.4125 13.5 18 13.5C17.5875 13.5 17.25 13.1625 17.25 12.75V10.5C17.25 10.0875 17.5875 9.74999 18 9.74999C18.4125 9.74999 18.75 10.0875 18.75 10.5V12.75Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export const PostItIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      d="M8.25 9C8.25 8.80109 8.32902 8.61032 8.46967 8.46967C8.61032 8.32902 8.80109 8.25 9 8.25H15C15.1989 8.25 15.3897 8.32902 15.5303 8.46967C15.671 8.61032 15.75 8.80109 15.75 9C15.75 9.19891 15.671 9.38968 15.5303 9.53033C15.3897 9.67098 15.1989 9.75 15 9.75H9C8.80109 9.75 8.61032 9.67098 8.46967 9.53033C8.32902 9.38968 8.25 9.19891 8.25 9ZM9 12.75H15C15.1989 12.75 15.3897 12.671 15.5303 12.5303C15.671 12.3897 15.75 12.1989 15.75 12C15.75 11.8011 15.671 11.6103 15.5303 11.4697C15.3897 11.329 15.1989 11.25 15 11.25H9C8.80109 11.25 8.61032 11.329 8.46967 11.4697C8.32902 11.6103 8.25 11.8011 8.25 12C8.25 12.1989 8.32902 12.3897 8.46967 12.5303C8.61032 12.671 8.80109 12.75 9 12.75ZM12 14.25H9C8.80109 14.25 8.61032 14.329 8.46967 14.4697C8.32902 14.6103 8.25 14.8011 8.25 15C8.25 15.1989 8.32902 15.3897 8.46967 15.5303C8.61032 15.671 8.80109 15.75 9 15.75H12C12.1989 15.75 12.3897 15.671 12.5303 15.5303C12.671 15.3897 12.75 15.1989 12.75 15C12.75 14.8011 12.671 14.6103 12.5303 14.4697C12.3897 14.329 12.1989 14.25 12 14.25ZM21 4.5V14.6897C21.0006 14.8867 20.9621 15.082 20.8866 15.264C20.8111 15.446 20.7002 15.6112 20.5603 15.75L15.75 20.5603C15.6112 20.7002 15.446 20.8111 15.264 20.8866C15.082 20.9621 14.8867 21.0006 14.6897 21H4.5C4.10218 21 3.72064 20.842 3.43934 20.5607C3.15804 20.2794 3 19.8978 3 19.5V4.5C3 4.10218 3.15804 3.72064 3.43934 3.43934C3.72064 3.15804 4.10218 3 4.5 3H19.5C19.8978 3 20.2794 3.15804 20.5607 3.43934C20.842 3.72064 21 4.10218 21 4.5ZM4.5 19.5H14.25V15C14.25 14.8011 14.329 14.6103 14.4697 14.4697C14.6103 14.329 14.8011 14.25 15 14.25H19.5V4.5H4.5V19.5ZM15.75 15.75V18.4406L18.4397 15.75H15.75Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export const VectorIcon = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width="100%"
-    height="100%"
-    viewBox="0 0 20 22"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      d="M7.33339 2.06044C8.39113 -0.241682 11.6024 -0.242599 12.6615 2.05892L19.7094 17.3753C20.9575 20.0876 17.9633 22.784 15.4771 21.1867L11.5716 18.6775C10.6126 18.0614 9.39155 18.0617 8.43284 18.6784L4.52876 21.1898C2.04344 22.7886 -0.952308 20.0939 0.294227 17.3809L7.33339 2.06044Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-export const ShipIt = (props: React.SVGProps<SVGSVGElement>) => (
-  <svg
-    width={200}
-    height={138}
-    viewBox="0 0 200 138"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <g clipPath="url(#clip0_58_395)">
+export function LinkedInIcon(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg
+      width={30}
+      height={31}
+      viewBox="0 0 30 31"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
       <path
-        d="M196.347 -0.000488281H3.4011C2.49163 0.00838337 1.62282 0.377607 0.985215 1.0262C0.347615 1.67479 -0.00670438 2.5498 -2.94419e-05 3.45929V134.547C-0.00491606 135.455 0.350254 136.328 0.987713 136.975C1.62517 137.622 2.49297 137.99 3.4011 137.998H196.347C197.255 137.99 198.123 137.622 198.76 136.975C199.398 136.328 199.753 135.455 199.748 134.547V3.45959C199.755 2.55016 199.401 1.6752 198.763 1.02662C198.125 0.378031 197.256 0.00846372 196.347 -0.000488281Z"
-        fill="black"
-        fillOpacity={0.25}
+        d="M23.75 4.25C24.413 4.25 25.0489 4.51339 25.5178 4.98223C25.9866 5.45107 26.25 6.08696 26.25 6.75V24.25C26.25 24.913 25.9866 25.5489 25.5178 26.0178C25.0489 26.4866 24.413 26.75 23.75 26.75H6.25C5.58696 26.75 4.95107 26.4866 4.48223 26.0178C4.01339 25.5489 3.75 24.913 3.75 24.25V6.75C3.75 6.08696 4.01339 5.45107 4.48223 4.98223C4.95107 4.51339 5.58696 4.25 6.25 4.25H23.75ZM23.125 23.625V17C23.125 15.9192 22.6957 14.8828 21.9315 14.1185C21.1672 13.3543 20.1308 12.925 19.05 12.925C17.9875 12.925 16.75 13.575 16.15 14.55V13.1625H12.6625V23.625H16.15V17.4625C16.15 16.5 16.925 15.7125 17.8875 15.7125C18.3516 15.7125 18.7967 15.8969 19.1249 16.2251C19.4531 16.5533 19.6375 16.9984 19.6375 17.4625V23.625H23.125ZM8.6 11.2C9.15695 11.2 9.6911 10.9788 10.0849 10.5849C10.4788 10.1911 10.7 9.65695 10.7 9.1C10.7 7.9375 9.7625 6.9875 8.6 6.9875C8.03973 6.9875 7.50241 7.21007 7.10624 7.60624C6.71007 8.00241 6.4875 8.53973 6.4875 9.1C6.4875 10.2625 7.4375 11.2 8.6 11.2ZM10.3375 23.625V13.1625H6.875V23.625H10.3375Z"
+        fill="currentColor"
       />
-      <path
-        d="M168.045 0.685059H3.58439C2.81499 0.686757 2.07767 0.993572 1.53419 1.5382C0.990711 2.08282 0.685443 2.82078 0.685364 3.59019V113.656C0.686656 114.425 0.992514 115.161 1.53591 115.705C2.0793 116.248 2.81592 116.554 3.58439 116.555H168.045C168.813 116.554 169.55 116.248 170.093 115.705C170.636 115.161 170.942 114.425 170.944 113.656V3.59141C170.944 2.82206 170.638 2.08414 170.095 1.53953C169.552 0.994909 168.814 0.688059 168.045 0.68628V0.685059Z"
-        fill="#090814"
-      />
-      <path
-        d="M190.645 0.685059H3.60492C2.72962 0.686596 1.89069 1.03537 1.27227 1.65481C0.653857 2.27426 0.306476 3.11377 0.306396 3.98907V129.167C0.307528 130.041 0.655418 130.88 1.27376 131.498C1.89211 132.116 2.73045 132.464 3.60492 132.465H190.645C191.519 132.464 192.358 132.116 192.976 131.498C193.594 130.88 193.942 130.041 193.943 129.167V3.98846C193.943 3.11345 193.596 2.27417 192.978 1.65476C192.36 1.03534 191.522 0.686389 190.647 0.684448L190.645 0.685059Z"
-        fill="white"
-      />
-      <path
-        d="M184.983 17.9326H9.97101C9.7864 17.9326 9.60936 17.8593 9.47882 17.7288C9.34829 17.5982 9.27496 17.4212 9.27496 17.2366C9.27496 17.052 9.34829 16.8749 9.47882 16.7444C9.60936 16.6139 9.7864 16.5405 9.97101 16.5405H184.984C185.168 16.5405 185.345 16.6139 185.476 16.7444C185.606 16.8749 185.68 17.052 185.68 17.2366C185.68 17.4212 185.606 17.5982 185.476 17.7288C185.345 17.8593 185.168 17.9326 184.983 17.9326Z"
-        fill="#CACACA"
-      />
-      <path
-        d="M13.3694 12.9867C15.6221 12.9867 17.4482 11.1199 17.4482 8.81709C17.4482 6.5143 15.6221 4.64752 13.3694 4.64752C11.1167 4.64752 9.29053 6.5143 9.29053 8.81709C9.29053 11.1199 11.1167 12.9867 13.3694 12.9867Z"
-        fill="#3F3D56"
-      />
-      <path
-        d="M27.4592 12.9867C29.7119 12.9867 31.5381 11.1199 31.5381 8.81709C31.5381 6.5143 29.7119 4.64752 27.4592 4.64752C25.2065 4.64752 23.3804 6.5143 23.3804 8.81709C23.3804 11.1199 25.2065 12.9867 27.4592 12.9867Z"
-        fill="#3F3D56"
-      />
-      <path
-        d="M41.5494 12.9867C43.8021 12.9867 45.6282 11.1199 45.6282 8.81709C45.6282 6.5143 43.8021 4.64752 41.5494 4.64752C39.2967 4.64752 37.4705 6.5143 37.4705 8.81709C37.4705 11.1199 39.2967 12.9867 41.5494 12.9867Z"
-        fill="#059669"
-      />
-      <path
-        d="M181.586 12.9867C183.838 12.9867 185.665 11.1199 185.665 8.81709C185.665 6.5143 183.838 4.64752 181.586 4.64752C179.333 4.64752 177.507 6.5143 177.507 8.81709C177.507 11.1199 179.333 12.9867 181.586 12.9867Z"
-        fill="#E6E6E6"
-      />
-      <path
-        d="M147.084 49.7745H106.908C102.86 49.7745 99.5781 53.0563 99.5781 57.1046V66.4354C99.5781 70.4836 102.86 73.7654 106.908 73.7654H147.084C151.133 73.7654 154.414 70.4836 154.414 66.4354V57.1046C154.414 53.0563 151.133 49.7745 147.084 49.7745Z"
-        fill="#059669"
-      />
-      <path
-        d="M87.213 49.7745H47.0368C42.9886 49.7745 39.7068 53.0563 39.7068 57.1046V66.4354C39.7068 70.4836 42.9886 73.7654 47.0368 73.7654H87.213C91.2613 73.7654 94.543 70.4836 94.543 66.4354V57.1046C94.543 53.0563 91.2613 49.7745 87.213 49.7745Z"
-        fill="#E6E6E6"
-      />
-      <path
-        d="M126.311 68.2829C126.143 68.2829 125.981 68.2214 125.856 68.11C125.73 67.9986 125.65 67.845 125.63 67.6784C125.502 66.6094 125.019 65.6143 124.258 64.8518C123.499 64.095 122.51 63.6128 121.446 63.4817C121.33 63.4704 121.22 63.43 121.124 63.3643C121.028 63.2986 120.951 63.2097 120.899 63.1059C120.847 63.0022 120.823 62.887 120.827 62.7711C120.832 62.6552 120.866 62.5424 120.927 62.4433C121.296 61.8074 121.806 61.265 122.418 60.8577C123.03 60.4505 123.728 60.1894 124.457 60.0946C124.96 58.9184 125.781 57.9063 126.829 57.1728C127.877 56.4393 129.11 56.0137 130.387 55.944C130.399 55.944 130.412 55.944 130.424 55.944C131.151 55.9447 131.848 56.2338 132.362 56.7478C132.876 57.2618 133.165 57.9587 133.166 58.6857C133.166 58.6981 133.166 58.7105 133.166 58.723C133.096 60.0002 132.671 61.2325 131.937 62.2804C131.203 63.3283 130.191 64.15 129.015 64.6527C128.92 65.3844 128.657 66.0843 128.248 66.698C127.838 67.3117 127.293 67.8227 126.653 68.1912C126.549 68.2513 126.431 68.2829 126.311 68.2829ZM122.721 62.3608C123.667 62.6596 124.527 63.1811 125.229 63.8815C125.929 64.5838 126.45 65.4438 126.748 66.3893C127.325 65.7818 127.657 64.9833 127.682 64.1463C127.687 64.0082 127.733 63.8748 127.815 63.7634C127.896 63.6521 128.01 63.568 128.14 63.5223C129.161 63.1654 130.053 62.5137 130.704 61.6497C131.354 60.7857 131.734 59.7481 131.795 58.6683C131.79 58.3109 131.645 57.9696 131.393 57.7169C131.14 57.4641 130.799 57.32 130.441 57.315C129.361 57.3756 128.324 57.7552 127.459 58.4057C126.595 59.0563 125.943 59.9485 125.587 60.9696C125.541 61.1 125.457 61.2135 125.345 61.2952C125.234 61.377 125.101 61.4232 124.963 61.4277C124.126 61.4528 123.328 61.785 122.721 62.3608Z"
-        fill="white"
-      />
-      <path
-        d="M121.487 68.6201C121.251 68.6201 121.015 68.6058 120.781 68.5773C120.616 68.5614 120.461 68.4916 120.341 68.3789C120.22 68.2661 120.14 68.1166 120.112 67.9536C119.972 66.9498 120.157 65.9274 120.64 65.0363C121.138 64.1188 121.9 63.3719 122.827 62.8923C123.006 62.7973 123.213 62.7728 123.409 62.8237C123.605 62.8746 123.774 62.9972 123.884 63.1671C123.933 63.2501 123.965 63.3425 123.977 63.4385C123.989 63.5344 123.981 63.6318 123.953 63.7245C123.926 63.8171 123.879 63.903 123.817 63.9767C123.754 64.0503 123.677 64.1102 123.59 64.1524C123.023 64.436 122.542 64.8652 122.196 65.396C121.85 65.9268 121.651 66.5401 121.62 67.173C122.27 67.1609 122.907 66.9866 123.473 66.6659C124.039 66.3452 124.516 65.8883 124.86 65.3365C124.97 65.1666 125.139 65.044 125.335 64.9931C125.531 64.9422 125.738 64.9666 125.917 65.0616C126.004 65.1039 126.081 65.1637 126.144 65.2374C126.206 65.3111 126.253 65.397 126.28 65.4896C126.308 65.5823 126.316 65.6796 126.304 65.7756C126.292 65.8715 126.26 65.9639 126.211 66.0469C125.664 66.94 124.863 67.6499 123.911 68.0859C123.152 68.4389 122.324 68.6213 121.487 68.6201Z"
-        fill="white"
-      />
-      <path
-        d="M129.063 61.1219C128.851 61.1219 128.643 61.0588 128.466 60.9406C128.289 60.8223 128.151 60.6542 128.069 60.4575C127.988 60.2609 127.966 60.0445 128.008 59.8357C128.049 59.6269 128.152 59.4351 128.302 59.2846C128.453 59.1341 128.645 59.0316 128.854 58.99C129.062 58.9485 129.279 58.9698 129.475 59.0513C129.672 59.1328 129.84 59.2707 129.958 59.4477C130.077 59.6247 130.14 59.8328 130.14 60.0457C130.139 60.331 130.026 60.6046 129.824 60.8063C129.622 61.0081 129.349 61.1216 129.063 61.1219Z"
-        fill="white"
-      />
-      <path
-        d="M73.2404 99.9013H121.491C121.909 99.9013 122.309 99.7355 122.604 99.4403C122.899 99.1452 123.065 98.7449 123.065 98.3275C123.065 97.9101 122.899 97.5098 122.604 97.2146C122.309 96.9195 121.909 96.7537 121.491 96.7537H73.2404C72.823 96.7537 72.4227 96.9195 72.1276 97.2146C71.8324 97.5098 71.6666 97.9101 71.6666 98.3275C71.6666 98.7449 71.8324 99.1452 72.1276 99.4403C72.4227 99.7355 72.823 99.9013 73.2404 99.9013Z"
-        fill="#E4E4E4"
-      />
-      <path
-        d="M73.2408 107.838H94.1561C94.5735 107.838 94.9738 107.673 95.269 107.377C95.5641 107.082 95.7299 106.682 95.7299 106.265C95.7299 105.847 95.5641 105.447 95.269 105.152C94.9738 104.857 94.5735 104.691 94.1561 104.691H73.2408C72.8233 104.691 72.423 104.857 72.1278 105.152C71.8326 105.447 71.6668 105.847 71.6668 106.265C71.6668 106.682 71.8326 107.083 72.1278 107.378C72.423 107.673 72.8233 107.839 73.2408 107.839V107.838Z"
-        fill="#E4E4E4"
-      />
-    </g>
-    <defs>
-      <clipPath id="clip0_58_395">
-        <rect width={199.748} height={138} fill="white" />
-      </clipPath>
-    </defs>
-  </svg>
-);
+    </svg>
+  );
+}
