+++ src/components/notion/mobile-workspace-topbar.tsx
@@ -1,17 +1,10 @@
 "use client";
 
-import { GithubIcon, LinkedInIcon, XIcon } from "@/components/icons";
+import { NotionSidebar } from "@/components/notion/sidebar";
 import { cn } from "@/lib/utils";
-import { Check, Lock, Mail, Menu, Plus, Share2, Star, X } from "lucide-react";
-import Image from "next/image";
+import { Check, Lock, Menu, Share2, Star, X } from "lucide-react";
 import Link from "next/link";
-import { type ReactNode, useEffect, useState } from "react";
-
-type NavItem = {
-  href: string;
-  icon: string;
-  label: string;
-};
+import { useEffect, useState } from "react";
 
 type MobileWorkspaceTopbarProps = {
   activePath: string;
@@ -19,9 +12,7 @@ type MobileWorkspaceTopbarProps = {
   editedCommitTitle: string;
   editedCommitUrl: string;
   editedDate: string;
-  favorites: NavItem[];
   isFavorite: boolean;
-  musings: NavItem[];
 };
 
 export function MobileWorkspaceTopbar({
@@ -30,9 +21,7 @@ export function MobileWorkspaceTopbar({
   editedCommitTitle,
   editedCommitUrl,
   editedDate,
-  favorites,
   isFavorite,
-  musings,
 }: MobileWorkspaceTopbarProps) {
   const [copied, setCopied] = useState(false);
   const [open, setOpen] = useState(false);
@@ -126,22 +115,10 @@ export function MobileWorkspaceTopbar({
             className="absolute inset-0 bg-black/45"
             onClick={() => setOpen(false)}
           />
-          <aside className="relative flex h-full w-full flex-col border-r border-[#2b2b2b] bg-[#202020] text-[#b8b8b5] sm:w-[92vw] sm:max-w-[280px]">
-            <div className="flex h-[45px] items-center justify-between px-3">
-              <Link
-                href="/"
-                className="flex min-w-0 items-center gap-2.5 text-[14px] font-semibold text-[#f1f1ef]"
-                onClick={() => setOpen(false)}
-              >
-                <Image
-                  src="/img/notion.png"
-                  alt="Notion"
-                  width={24}
-                  height={24}
-                  className="h-6 w-6 rounded-sm"
-                />
-                <span className="truncate">brennen’s notion</span>
-              </Link>
+          <NotionSidebar
+            activePath={activePath}
+            className="relative w-full sm:w-[92vw] sm:max-w-[280px]"
+            headerAction={
               <button
                 type="button"
                 aria-label="Close sidebar"
@@ -150,124 +127,11 @@ export function MobileWorkspaceTopbar({
               >
                 <X className="h-5 w-5" />
               </button>
-            </div>
-
-            <nav className="flex flex-1 flex-col gap-9 px-2 pt-10 text-[14px] font-semibold">
-              <MobileSidebarGroup title="Favorites">
-                {favorites.map((item) => (
-                  <MobileSidebarLink
-                    key={item.href}
-                    active={activePath === item.href}
-                    item={item}
-                    onNavigate={() => setOpen(false)}
-                  />
-                ))}
-              </MobileSidebarGroup>
-
-              <MobileSidebarGroup
-                action={
-                  <button aria-label="Add musing" className="rounded p-1">
-                    <Plus className="h-4 w-4" />
-                  </button>
-                }
-                title="Musings"
-              >
-                {musings.map((item) => (
-                  <MobileSidebarLink
-                    key={item.href}
-                    active={activePath === item.href}
-                    item={item}
-                    onNavigate={() => setOpen(false)}
-                  />
-                ))}
-              </MobileSidebarGroup>
-            </nav>
-
-            <div className="border-t border-[#303030] px-4 py-3">
-              <div className="flex items-center gap-3.5 text-[#c7c7c5]">
-                <MobileSocialLink href="mailto:web@brennen.dev" label="Email">
-                  <Mail className="h-4.5 w-4.5" />
-                </MobileSocialLink>
-                <MobileSocialLink
-                  href="https://www.linkedin.com/in/brennenho/"
-                  label="LinkedIn"
-                >
-                  <LinkedInIcon className="h-4.5 w-4.5" />
-                </MobileSocialLink>
-                <MobileSocialLink
-                  href="https://github.com/brennenho"
-                  label="GitHub"
-                >
-                  <GithubIcon className="h-4.5 w-4.5" />
-                </MobileSocialLink>
-                <MobileSocialLink href="https://x.com/brennenho_" label="X">
-                  <XIcon className="h-4.5 w-4.5" />
-                </MobileSocialLink>
-              </div>
-            </div>
-          </aside>
+            }
+            onNavigate={() => setOpen(false)}
+          />
         </div>
       ) : null}
     </>
-  );
-}
-
-function MobileSidebarGroup({
-  action,
-  children,
-  title,
-}: {
-  action?: ReactNode;
-  children: ReactNode;
-  title: string;
-}) {
-  return (
-    <div>
-      <div className="mb-1.5 flex items-center justify-between px-2 text-[12px] font-semibold text-[#9b9b98]">
-        <span>{title}</span>
-        {action}
-      </div>
-      <div className="flex flex-col gap-1">{children}</div>
-    </div>
-  );
-}
-
-function MobileSidebarLink({
-  active,
-  item,
-  onNavigate,
-}: {
-  active: boolean;
-  item: NavItem;
-  onNavigate: () => void;
-}) {
-  return (
-    <Link
-      href={item.href}
-      onClick={onNavigate}
-      className={cn(
-        "flex h-8 items-center gap-2 rounded-sm px-2 text-[#b8b8b5] transition-colors hover:bg-[#30302f]",
-        active && "bg-[#373736] text-[#f1f1ef]",
-      )}
-    >
-      <span className="w-5 text-[18px] leading-none">{item.icon}</span>
-      <span className="truncate">{item.label}</span>
-    </Link>
-  );
-}
-
-function MobileSocialLink({
-  children,
-  href,
-  label,
-}: {
-  children: ReactNode;
-  href: string;
-  label: string;
-}) {
-  return (
-    <Link href={href} target="_blank" rel="noreferrer" aria-label={label}>
-      {children}
-    </Link>
   );
 }
