+++ src/components/notion/workspace-shell.tsx
@@ -1,5 +1,5 @@
-import { GithubIcon, LinkedInIcon, XIcon } from "@/components/icons";
 import { MobileWorkspaceTopbar } from "@/components/notion/mobile-workspace-topbar";
+import { isFavoritePath, NotionSidebar } from "@/components/notion/sidebar";
 import { TopbarActions } from "@/components/notion/topbar-actions";
 import { cn } from "@/lib/utils";
 import {
@@ -7,11 +7,8 @@ import {
   ChevronRight,
   Flower2,
   Lock,
-  Mail,
-  Plus,
   Sprout,
 } from "lucide-react";
-import Image from "next/image";
 import Link from "next/link";
 import type { ReactNode } from "react";
 
@@ -30,32 +27,6 @@ const pageTitleClassName =
 const bodyTextClassName =
   "text-[16px] leading-[1.5] font-medium text-[#f1f1ef]";
 
-const favorites = [
-  { href: "/", icon: "👋", label: "hey, i’m brennen" },
-  { href: "/work", icon: "💼", label: "work" },
-  { href: "/nursery", icon: "🌱", label: "nursery" },
-  { href: "/playground", icon: "🛝", label: "playground" },
-];
-
-const musings = [
-  {
-    href: "/musings/summer-in-the-city",
-    icon: "🌉",
-    label: "summer in the city - jun ’26",
-  },
-  {
-    href: "/musings/side-quests",
-    icon: "🕵️",
-    label: "side quests - apr ’26",
-  },
-  { href: "/musings/watchlist", icon: "👀", label: "watchlist - feb ’26" },
-  {
-    href: "/musings/new-beginnings",
-    icon: "🌻",
-    label: "new beginnings - jan ’26",
-  },
-];
-
 export function WorkspaceShell({
   children,
   editedCommitTimestamp,
@@ -64,7 +35,7 @@ export function WorkspaceShell({
   editedDate,
   activePath = "/",
 }: WorkspaceShellProps) {
-  const isFavorite = favorites.some((item) => item.href === activePath);
+  const isFavorite = isFavoritePath(activePath);
 
   return (
     <div className="min-h-screen bg-[#191919] text-[#f1f1ef]">
@@ -76,9 +47,7 @@ export function WorkspaceShell({
           editedCommitTitle={editedCommitTitle}
           editedCommitUrl={editedCommitUrl}
           editedDate={editedDate}
-          favorites={favorites}
           isFavorite={isFavorite}
-          musings={musings}
         />
         <WorkspaceTopbar
           editedCommitTimestamp={editedCommitTimestamp}
@@ -158,137 +127,10 @@ function WorkspaceTopbar({
 
 function WorkspaceSidebar({ activePath }: { activePath: string }) {
   return (
-    <aside className="fixed inset-y-0 left-0 z-50 hidden w-[244px] flex-col border-r border-[#2b2b2b] bg-[#202020] text-[#b8b8b5] min-[900px]:flex">
-      <div className="flex h-[45px] items-center justify-between px-3">
-        <Link
-          href="/"
-          className="flex min-w-0 items-center gap-2.5 text-[14px] font-semibold text-[#f1f1ef]"
-        >
-          <Image
-            src="/img/notion.png"
-            alt="Notion"
-            width={24}
-            height={24}
-            className="h-6 w-6 rounded-sm"
-          />
-          <span className="truncate">brennen’s notion</span>
-        </Link>
-      </div>
-
-      <nav className="flex flex-1 flex-col gap-9 px-2 pt-10 text-[14px] font-semibold">
-        <SidebarGroup title="Favorites">
-          {favorites.map((item) => (
-            <SidebarLink
-              key={item.href}
-              active={activePath === item.href}
-              href={item.href}
-              icon={item.icon}
-              label={item.label}
-            />
-          ))}
-        </SidebarGroup>
-
-        <SidebarGroup
-          action={
-            <button aria-label="Add musing" className="rounded p-1">
-              <Plus className="h-4 w-4" />
-            </button>
-          }
-          title="Musings"
-        >
-          {musings.map((item) => (
-            <SidebarLink
-              key={item.href}
-              active={activePath === item.href}
-              href={item.href}
-              icon={item.icon}
-              label={item.label}
-            />
-          ))}
-        </SidebarGroup>
-      </nav>
-
-      <div className="border-t border-[#303030] px-4 py-3">
-        <div className="flex items-center gap-3.5 text-[#c7c7c5]">
-          <SocialLink href="mailto:web@brennen.dev" label="Email">
-            <Mail className="h-4.5 w-4.5" />
-          </SocialLink>
-          <SocialLink
-            href="https://www.linkedin.com/in/brennenho/"
-            label="LinkedIn"
-          >
-            <LinkedInIcon className="h-4.5 w-4.5" />
-          </SocialLink>
-          <SocialLink href="https://github.com/brennenho" label="GitHub">
-            <GithubIcon className="h-4.5 w-4.5" />
-          </SocialLink>
-          <SocialLink href="https://x.com/brennenho_" label="X">
-            <XIcon className="h-4.5 w-4.5" />
-          </SocialLink>
-        </div>
-      </div>
-    </aside>
-  );
-}
-
-function SidebarGroup({
-  action,
-  children,
-  title,
-}: {
-  action?: ReactNode;
-  children: ReactNode;
-  title: string;
-}) {
-  return (
-    <div>
-      <div className="mb-1.5 flex items-center justify-between px-2 text-[12px] font-semibold text-[#9b9b98]">
-        <span>{title}</span>
-        {action}
-      </div>
-      <div className="flex flex-col gap-1">{children}</div>
-    </div>
-  );
-}
-
-function SidebarLink({
-  active,
-  href,
-  icon,
-  label,
-}: {
-  active: boolean;
-  href: string;
-  icon: string;
-  label: string;
-}) {
-  return (
-    <Link
-      href={href}
-      className={cn(
-        "flex h-8 items-center gap-2 rounded-sm px-2 text-[#b8b8b5] transition-colors hover:bg-[#30302f]",
-        active && "bg-[#373736] text-[#f1f1ef]",
-      )}
-    >
-      <span className="w-5 text-[18px] leading-none">{icon}</span>
-      <span className="truncate">{label}</span>
-    </Link>
-  );
-}
-
-function SocialLink({
-  children,
-  href,
-  label,
-}: {
-  children: ReactNode;
-  href: string;
-  label: string;
-}) {
-  return (
-    <Link href={href} target="_blank" rel="noreferrer" aria-label={label}>
-      {children}
-    </Link>
+    <NotionSidebar
+      activePath={activePath}
+      className="fixed inset-y-0 left-0 z-50 hidden w-[244px] min-[900px]:flex"
+    />
   );
 }
