--- src/components/scroll-animation.tsx
@@ -1,40 +1,0 @@
-"use client";
-
-import { cn } from "@/lib/utils";
-import { motion, useInView } from "framer-motion";
-import { useRef } from "react";
-
-const item = {
-  hidden: { opacity: 0, y: 20, filter: "blur(10px)" },
-  visible: {
-    opacity: 1,
-    y: 0,
-    filter: "blur(0px)",
-    transition: {
-      duration: 0.6,
-      ease: [0.17, 0.67, 0.83, 0.67] as const,
-    },
-  },
-};
-
-interface ScrollAnimationProps {
-  children: React.ReactNode;
-  className?: string;
-}
-
-export function ScrollAnimation({ children, className }: ScrollAnimationProps) {
-  const ref = useRef(null);
-  const isInView = useInView(ref, { once: true, amount: 0.3 });
-
-  return (
-    <motion.div
-      ref={ref}
-      initial="hidden"
-      animate={isInView ? "visible" : "hidden"}
-      variants={item}
-      className={cn("w-full", className)}
-    >
-      {children}
-    </motion.div>
-  );
-}
