+++ src/components/typography.tsx
@@ -1,9 +1,3 @@
-import { ScrollAnimation } from "@/components/scroll-animation";
-
 export function Heading({ children }: { children: React.ReactNode }) {
-  return (
-    <ScrollAnimation className="text-2xl leading-none font-bold">
-      {children}
-    </ScrollAnimation>
-  );
+  return <h1 className="text-2xl leading-none font-bold">{children}</h1>;
 }
