+++ src/lib/supabase/server.ts
@@ -24,9 +24,7 @@ export async function createClient() {
               cookieStore.set(name, value, options),
             );
           } catch {
-            // The `setAll` method was called from a Server Component.
-            // This can be ignored if you have middleware refreshing
-            // user sessions.
+            // Server Components cannot set cookies directly.
           }
         },
       },
